user sorting

+++ app/Http/Controllers/UserController.php
@@ -35,10 +35,16 @@ class UserController extends Controller
     public function destroy($userId)
     {
         $user = User::findOrFail($userId);
-        $user->delete();
         $user['role'] = $user->getRoleNames()->first();
-        return response()->json($user);
-
+        if ($user['role'] = 'Super Admin')
+        {
+            abort('401');
+        }
+        else {
+            $user->delete();
+            return response()->json($user);
+    
+        }
     }
 
     public function multidestroy(Request $request)
@@ -48,9 +54,38 @@ class UserController extends Controller
         foreach ($idArray as $id)
         {
             $user = User::findorFail($id);
-            $user->delete();
+            $user['role'] = $user->getRoleNames()->first();
+
+            if ($user['role'] != 'Super Admin')
+            {
+                $user->delete();
+            }
         }
 
         return response()->json($idArray);
+    }
+
+    public function create(Request $request)
+    {
+      $validator = Validator::make($request->all(), [
+          'name'  => 'required|max:255',
+          'email' => 'required|email|unique:users',
+        ]);
+        // If validator fails, short circut and redirect with errors
+        if($validator->fails()){
+          return back()
+            ->withErrors($validator)
+            ->withInput();
+        }
+        //generate a password for the new users
+        $pw = User::generatePassword();
+        //add new user to database
+        $user = new User;
+        $user->name = $request->input('name');
+        $user->email = $request->input('email');
+        $user->password = $pw;
+        $user->assignRole('Gebruiker');
+        $user->save();
+        User::sendWelcomeEmail($user);
     }
 }
